Changes to user access.

+++ app/helpers/application_helper.rb
@@ -12,7 +12,7 @@ module ApplicationHelper
 
     # Generate the dynamic path helper (e.g., admin_blog_posts_path)
     path_helper = "admin_#{model_path_name}_path"
-    path = send(path_helper, sort: column, direction: current_direction)
+    path        = send(path_helper, sort: column, direction: current_direction)
 
     # Generate the sortable link with the arrow
     link_to "#{display_name} #{arrow}".html_safe, path, class: custom_class
@@ -20,22 +20,25 @@ module ApplicationHelper
 
   def action_links(resource, show_path, edit_path, delete_path = nil)
     delete_path ||= resource
-    content_tag(:div, class: "action-links") do
-      [
-        link_to("View",
-                show_path,
-                class: "btn-link me-2"),
-        link_to("Edit",
-                edit_path,
-                class: "btn-link me-2"),
-        link_to("Delete",
-                delete_path,
-                method: :delete,
-                data: { confirm: "Are you sure?" },
-                class: "btn-link",
-                style: "color: red",
-                onclick: "if (!confirm('Are you sure you want to delete this?')) { return false; }")
-      ].join.html_safe
+    links       = [
+      link_to("View",
+              show_path,
+              class: "btn-link me-2"),
+      link_to("Edit",
+              edit_path,
+              class: "btn-link me-2"),
+    ]
+
+    if @application_user.admin?
+      links << link_to("Delete",
+                       delete_path,
+                       method:  :delete,
+                       data:    { confirm: "Are you sure?" },
+                       class:   "btn-link",
+                       style:   "color: red",
+                       onclick: "if (!confirm('Are you sure you want to delete this?')) { return false; }")
     end
+
+    content_tag(:div, class: "action-links") { links.join.html_safe }
   end
 end
